notMNIST beginner code

+++ notMNIST/read_notMNIST.py
@@ -1,5 +1,6 @@
 __author__ = 'Charlie'
 # Converting udacity tensorflow tutorial code such that notMNIST can be accessed with ease
+#Note that the taining data set size is set to 20k - The data set has 500k in large and 20k in small
 import numpy as np
 import os
 import sys
@@ -190,22 +191,22 @@ def initialize_read(data_dir):
     valid_dataset, valid_labels = randomize(valid_dataset, valid_labels)
 
     pickle_file = os.path.join(data_dir, 'notMNIST.pickle')
-
-    try:
-        f = open(pickle_file, 'wb')
-        save = {
-            'train_dataset': train_dataset,
-            'train_labels': train_labels,
-            'valid_dataset': valid_dataset,
-            'valid_labels': valid_labels,
-            'test_dataset': test_dataset,
-            'test_labels': test_labels,
-        }
-        pickle.dump(save, f, pickle.HIGHEST_PROTOCOL)
-        f.close()
-    except Exception as e:
-        print('Unable to save data to', pickle_file, ':', e)
-        raise
+    if not os.path.exists(pickle_file):
+        try:
+            f = open(pickle_file, 'wb')
+            save = {
+                'train_dataset': train_dataset,
+                'train_labels': train_labels,
+                'valid_dataset': valid_dataset,
+                'valid_labels': valid_labels,
+                'test_dataset': test_dataset,
+                'test_labels': test_labels,
+            }
+            pickle.dump(save, f, pickle.HIGHEST_PROTOCOL)
+            f.close()
+        except Exception as e:
+            print('Unable to save data to', pickle_file, ':', e)
+            raise
 
     statinfo = os.stat(pickle_file)
     print('Compressed pickle size:', statinfo.st_size)
